ADD min-max scaling to pink file creation
Processing the Pink file via

```sh
Pink --train data/P205+55_pink_v2.bin P205.som
```

works as far as creating a valid SOM-file, which can be output via
the PINK Python script show_som.py. Both SOM.pdf files under docs/soms
were created with an input of 1000 200x200 images and 5 iterations. The
first run used unscaled array data whereas the second run used min-max
scaled data. Both versions unfortunately contain big pink artefacts.

TODO: Find out how to construct "good looking" SOMs. Higher iteration,
denoising of images, centering the brightest point?

+++ hda_fits/fits.py
@@ -105,7 +105,7 @@ def create_cutout2D_as_flattened_numpy_array(
     size: Union[int, RectangleSize],
     wcs: WCS = None,
 ) -> np.ndarray:
-    return create_cutout2D(hdu, coordinates, size, wcs).data.flatten()
+    return create_cutout2D(hdu, coordinates, size, wcs).data.flatten().astype("float32")
 
 
 def create_cutout2D_as_updated_hdu(
